crowbar: Fix the bootstrap installation
On installation, the data bag configuration is created before the admin
server gets allocated an IP address. So we need to deal with the case
where the node may not have an admin network in the code.

This also means we will need to regenerate the data bag configuration at
some point; this will be done in the install script, after we move the
admin server to hardware-installing.

(cherry picked from commit 3963959d9716f83dbb82374df211f77b1dc4477e)